+++ __init__.py
@@ -19,9 +19,9 @@
 bl_info = {
     "name": "Path Actions",
     "author": "Samuel Bernou",
-    "version": (1, 7, 0),
+    "version": (1, 8, 0),
     "blender": (2, 80, 0),
-    "location": "Window top right corner and browser footer",
+    "location": "Window top right corner, browser footer, addon prefs",
     "description": "Open blend file folder in OS explorer",
     "warning": "",
     "wiki_url": "https://github.com/Pullusb/SB_path_actions",
@@ -31,12 +31,14 @@ bl_info = {
 from . import (
             operators,
             blend_open_ops,
+            open_addons_path,
             panels,
             keymap)
 
 import bpy, os
 import rna_keymap_ui
 from pathlib import Path
+# import addon_utils
 
 # def get_hotkey_entry_item(km, kmi_name):
 def get_hotkey_entry_item(km, kmi_name, kmi_value, properties):
@@ -52,6 +54,19 @@ class PATH_addon_preferences(bpy.types.AddonPreferences):
     """ Preference Settings Addon Panel"""
     bl_idname = __name__.split('.')[0] # __name__
     bl_label = "Addon Preferences"
+
+    # show_addon_loc_details : bpy.props.BoolProperty(name='All Addons load folders', default=False)
+
+    show_addon_open_buttons : bpy.props.BoolProperty(name='Addons Directory Shortcuts', default=False)
+
+    filter : bpy.props.EnumProperty(
+        name="Filter", description="Filter addon list to open their directory",
+        default='ALL', options={'HIDDEN'},
+        items=(
+            ('ALL', 'All', 'List all addons', 0),
+            ('ACTIVE', 'Enabled', 'List only enabled addons', 1),   
+            ('INACTIVE', 'Disabled', 'List only Disabled addons', 2),   
+            ))
 
     def draw(self, context):
         layout = self.layout
@@ -113,10 +128,52 @@ class PATH_addon_preferences(bpy.types.AddonPreferences):
         row.operator("wm.path_open", text='Config Folder').filepath = bpy.utils.user_resource('CONFIG')
         row.operator("wm.path_open", text='Application Folder').filepath = str(Path(bpy.app.binary_path).parent)
 
+        ## Addons external pathes
+        
+        # box = layout.box()
+        # box.prop(self, 'show_addon_loc_details')
+        # if self.show_addon_loc_details:
+        #     for fp in addon_utils.paths():
+        #         row = box.row()
+        #         # row.operator("wm.path_open", text=fp).filepath = fp
+        #         row.operator("path.open_browser", text=fp).filepath = fp
+        
+        
+        ## addon folder list
+        # layout.operator("path.open_addon_directory", text='Search And Open Addon Folder', icon='FILE_FOLDER')
+
+        box = layout.box()
+        expand_icon = 'TRIA_DOWN' if self.show_addon_open_buttons else 'TRIA_RIGHT'
+        row = box.row()
+        row.prop(self, 'show_addon_open_buttons', icon=expand_icon, emboss=False,) # text='Open Individual Addon Directory'
+        row.operator("path.open_addon_directory", text='Search', icon='VIEWZOOM')
+        if self.show_addon_open_buttons:
+            box.prop(self, 'filter')
+            ## responsive column list
+            if context.area.width > 1750 * context.preferences.view.ui_scale:
+                colnum = 6
+            elif context.area.width > 1450 * context.preferences.view.ui_scale:
+                colnum = 5
+            elif context.area.width > 1150 * context.preferences.view.ui_scale:
+                colnum = 4
+            elif context.area.width > 850 * context.preferences.view.ui_scale:
+                colnum = 3
+            elif context.area.width > 550 * context.preferences.view.ui_scale:
+                colnum = 2
+            else:
+                colnum = 1
+
+            cf = box.column_flow(columns=colnum, align=True)
+            for ad in open_addons_path.get_addons_modules_infos(self.filter):
+                # path, name, diskname
+                # col.operator("path.open_browser", text=ad[1]).filepath = ad[0]
+                cf.operator("path.open_browser", text=ad[1]).filepath = ad[0]
+
 
 def register():
     blend_open_ops.register()
     operators.register()
+    open_addons_path.register()
     panels.register()
     keymap.register()
     bpy.utils.register_class(PATH_addon_preferences)
@@ -125,6 +182,7 @@ def unregister():
     bpy.utils.unregister_class(PATH_addon_preferences)
     keymap.unregister()
     panels.unregister()
+    open_addons_path.unregister()
     operators.unregister()
     blend_open_ops.unregister()